% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_solutions.R
\name{install_solutions}
\alias{install_solutions}
\title{Install workshop solutions}
\usage{
install_solutions(destdir = getOption("usethis.destdir"))
}
\arguments{
\item{destdir}{The path on your computer where you would like the workshop
installed. By default, this will install somewhere conspicuous, like your
desktop, although you can tell `install_solutions()` exactly where you want
it to download.}
}
\description{
`install_solutions()` will download the solutions to the exercises for the workshop
Causal Inference in R on your computer. Then, it will open a new RStudio
Project containing the files you'll need.
}
