% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scripts.R
\name{list_workshop_scripts}
\alias{list_workshop_scripts}
\title{List Available Workshop Scripts}
\usage{
list_workshop_scripts()
}
\value{
Data frame with script information (invisibly)
}
\description{
Shows the workshop scripts included in the package with brief descriptions.
}
\examples{
list_workshop_scripts()

}
