% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_heterogeneity_distribution}
\alias{plot_heterogeneity_distribution}
\title{Plot Treatment Effect Heterogeneity Distribution}
\usage{
plot_heterogeneity_distribution(
  data,
  tau_var = "tau_hat",
  title = "Distribution of Individual Treatment Effects",
  bins = 30
)
}
\arguments{
\item{data}{A data frame with predicted treatment effects}

\item{tau_var}{Character. Name of treatment effect variable (default: "tau_hat")}

\item{title}{Character. Plot title}

\item{bins}{Integer. Number of histogram bins (default: 30)}
}
\value{
A ggplot object
}
\description{
Creates a histogram showing the distribution of predicted individual
treatment effects from a causal forest.
}
\examples{
cf_results <- workshop_causal_forest()
plot_heterogeneity_distribution(cf_results$predictions)

}
