% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_workshop.R
\name{install_workshop}
\alias{install_workshop}
\title{Install workshop materials}
\usage{
install_workshop(destdir = getOption("usethis.destdir"))
}
\arguments{
\item{destdir}{The path on your computer where you would like the workshop
installed. By default, this will install somewhere conspicous, like your
desktop, although you can tell `install_workshop()` exactly where you want
it to download.}
}
\description{
`install_course()` will install the workshop materials for the workshop
Causal Inference in R on your computer. Then, it will open a new RStudio
Project containing the files you'll need.
}
